+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## dev  (Unreleased)
+## 0.1.3
+
+FEATURES:
+* migration support with moved{} block from `aws_ssm_parameter` to `fastssm_parameter`
+
+FIXES:
+* properly handle `insecure_value` and `value` in the state
 
 ## 0.1.2
 
@@ -6,7 +12,7 @@ FEATURES:
 * validate AWS credentials before sending any API calls
 * update documentation
 
-## 0.1.1 (Unreleased)
+## 0.1.1
 
 FEATURES:
 * provision SSM parameters (resource and data_source)